# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UAT Test Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /test-cycles:
    get:
      operationId: listTestCycles
      tags: [testCycles]
      summary: List test cycles
      parameters:
        - name: releaseId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of test cycles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestCycle"
    post:
      operationId: createTestCycle
      tags: [testCycles]
      summary: Create a test cycle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestCycleBody"
      responses:
        "201":
          description: Created test cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCycle"
  /test-cycles/{id}:
    get:
      operationId: getTestCycle
      tags: [testCycles]
      summary: Get test cycle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test cycle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCycleDetail"
    patch:
      operationId: updateTestCycle
      tags: [testCycles]
      summary: Update test cycle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTestCycleBody"
      responses:
        "200":
          description: Updated test cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCycle"
...
  /test-executions:
    get:
      operationId: listTestExecutions
      tags: [testExecutions]
      summary: List test executions
      parameters:
        - name: testCaseId
          in: query
          required: false
          schema:
            type: integer
        - name: testCycleId
          in: query
          required: false
          schema:
            type: integer
        - name: executorId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of test executions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestExecution"
    post:
      operationId: createTestExecution
      tags: [testExecutions]
      summary: Create a test execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestExecutionBody"
      responses:
        "201":
          description: Created test execution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestExecution"
  /test-executions/{id}:
    get:
      operationId: getTestExecution
      tags: [testExecutions]
      summary: Get test execution by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test execution details with steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestExecutionDetail"
    patch:
      operationId: updateTestExecution
      tags: [testExecutions]
      summary: Update test execution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTestExecutionBody"
      responses:
        "200":
          description: Updated test execution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestExecution"
  /test-execution-steps/{id}:
    patch:
      operationId: updateTestExecutionStep
      tags: [testExecutions]
      summary: Update a test execution step result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTestExecutionStepBody"
      responses:
        "200":
          description: Updated step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestExecutionStep"
...
components:
  schemas:
...
    TestCycle:
      type: object
      properties:
        id:
          type: integer
        releaseId:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, releaseId, name, type, status, startDate, endDate, createdAt]
    TestCycleDetail:
      type: object
      properties:
        id:
          type: integer
        releaseId:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        testPlans:
          type: array
          items:
            $ref: "#/components/schemas/TestPlan"
        stats:
          $ref: "#/components/schemas/CycleStats"
      required: [id, releaseId, name, type, status, startDate, endDate, createdAt, testPlans, stats]
    CycleStats:
      type: object
      properties:
        totalPlans:
          type: integer
        totalCases:
          type: integer
        passed:
          type: integer
        failed:
          type: integer
        blocked:
          type: integer
        inProgress:
          type: integer
        notStarted:
          type: integer
        passRate:
          type: number
      required: [totalPlans, totalCases, passed, failed, blocked, inProgress, notStarted, passRate]
    CreateTestCycleBody:
      type: object
      properties:
        releaseId:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        ownerId:
          type: ["integer", "null"]
      required: [releaseId, name, type, startDate, endDate]
    UpdateTestCycleBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        ownerId:
          type: ["integer", "null"]
    TestPlan:
      type: object
      properties:
        id:
          type: integer
        testCycleId:
          type: integer
        name:
          type: string
        description:
          type: string
        module:
          type: string
        status:
          type: string
        entryCriteria:
          type: string
        exitCriteria:
          type: string
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, testCycleId, name, description, module, status, createdAt]
...
    TestCaseDetail:
      type: object
      properties:
        id:
          type: integer
        scenarioId:
          type: integer
        title:
          type: string
        description:
          type: string
        preconditions:
          type: string
        postconditions:
          type: string
        priority:
          type: string
        complexity:
          type: string
        status:
          type: string
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        estimatedMinutes:
          type: ["integer", "null"]
        lastExecutionStatus:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        steps:
          type: array
          items:
            $ref: "#/components/schemas/TestStep"
        executions:
          type: array
          items:
            $ref: "#/components/schemas/TestExecution"
      required: [id, scenarioId, title, description, priority, complexity, status, createdAt, steps, executions]
    TestStep:
      type: object
      properties:
        id:
          type: integer
        testCaseId:
          type: integer
        stepNumber:
          type: integer
        action:
          type: string
        expectedResult:
          type: string
        testData:
          type: string
      required: [id, testCaseId, stepNumber, action, expectedResult]
...
    TestExecution:
      type: object
      properties:
        id:
          type: integer
        testCaseId:
          type: integer
        testCycleId:
          type: integer
        executorId:
          type: ["integer", "null"]
        executorName:
          type: ["string", "null"]
        status:
          type: string
        result:
          type: string
        notes:
          type: string
        executedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        testCaseTitle:
          type: ["string", "null"]
      required: [id, testCaseId, testCycleId, status, result, createdAt]
    TestExecutionDetail:
      type: object
      properties:
        id:
          type: integer
        testCaseId:
          type: integer
        testCycleId:
          type: integer
        executorId:
          type: ["integer", "null"]
        executorName:
          type: ["string", "null"]
        status:
          type: string
        result:
          type: string
        notes:
          type: string
        executedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        testCaseTitle:
          type: ["string", "null"]
        testCase:
          $ref: "#/components/schemas/TestCaseDetail"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/TestExecutionStep"
        defects:
          type: array
          items:
            $ref: "#/components/schemas/Defect"
      required: [id, testCaseId, testCycleId, status, result, createdAt, testCase, steps, defects]
    TestExecutionStep:
      type: object
      properties:
        id:
          type: integer
        executionId:
          type: integer
        testStepId:
          type: integer
        stepNumber:
          type: integer
        action:
          type: string
        expectedResult:
          type: string
        actualResult:
          type: string
        status:
          type: string
        notes:
          type: string
      required: [id, executionId, testStepId, stepNumber, action, expectedResult, status]
    CreateTestExecutionBody:
      type: object
      properties:
        testCaseId:
          type: integer
        testCycleId:
          type: integer
        executorId:
          type: ["integer", "null"]
      required: [testCaseId, testCycleId]
    UpdateTestExecutionBody:
      type: object
      properties:
        status:
          type: string
        result:
          type: string
        notes:
          type: string
        executedAt:
          type: ["string", "null"]
          format: date-time
    UpdateTestExecutionStepBody:
      type: object
      properties:
        actualResult:
          type: string
        status:
          type: string
        notes:
          type: string
    Defect:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        severity:
          type: string
        priority:
          type: string
        status:
          type: string
        rootCause:
          type: string
        testExecutionId:
          type: ["integer", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        reporterId:
          type: ["integer", "null"]
        reporterName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, description, severity, priority, status, createdAt, updatedAt]
...