# lib/api-zod/src/generated/types/cycleStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UAT Test Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CycleStats {
  totalPlans: number;
  totalCases: number;
  passed: number;
  failed: number;
  blocked: number;
  inProgress: number;
  notStarted: number;
  passRate: number;
}