# lib/api-zod/src/generated/types/testExecutionStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UAT Test Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TestExecutionStep {
  id: number;
  executionId: number;
  testStepId: number;
  stepNumber: number;
  action: string;
  expectedResult: string;
  actualResult?: string;
  status: string;
  notes?: string;
}